Moved logging out of ROS topic file
